refactor(api): extract filesystem-tree ancestor helper into logic module
Keep the endpoint module to its default-exported handler per the backend convention; move the pure collectMissingAncestorIds helper (and its test) into read-filesystem-tree.logic.ts so the test no longer loads the handler's Elysia/DB dependencies.
